Enable the generic op block factor free dim check (#5208)
The core change of this commit is enabling the check of free block
factors (ones that don't participate in worker grid dims, like matmul
inner dim). This enforces that block factors are always in alignment
with the input tensor grid shapes and they can now be directly used for
loop nest creation.

This change also:
- Automatically infers block factors in the generic op defualt builder.
- Move `concatInversePermutationMap` to its own utility function since
it's used in multiple places.
- Introduces new method `getLegalBlockFactorFactorizations`, which can
be used by block factor / allocator analysis to choose legal factors
from.