Restrict D2M matmul builder test inputs and other improvements (#5293)
### Ticket
None

### Problem description
Currently the metal matmul tests are marked as failing, this could hide
all kinds of regressions except for the crashing ones.
Due to the importance of this op, we should try our best to make sure
it's properly validated in the CI.

### What's changed
Restrict the input range of the matmul tests to circumvent the f32 to
tf32 accuracy loss issue.
Also add a helper function to display the exact shapes in the tests'
names, so it's easier to guess why a test is failing.

### Checklist
- [x] New/Existing tests provide coverage for changes

Signed-off-by: wenbinlyuTT <wenbinlyu@tenstorrent.com>