[ttnn.jit] Run flatbuffers from memory (#5318)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/5055

### Problem description
We want to be able to run flatbuffers from memory instead of dumping to
a file and then running from the path of the file.

### What's changed
Wrap shared_ptr to flatbuffer binary in `nb::capsule` to expose it to
python land.

ty @nsmithtt 🙏 
### Checklist
- [x] New/Existing tests provide coverage for changes <- existing jit
tests will all run flatbuffer from memory now.