Add canonicalizers for ClampScalarOp (#5365)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-mlir/issues/5360)

### Problem description
Consecutive `clamp_scalar` operations can be simplified into a single
`clamp_scalar`.
Additionally, `clamp_scalar(0, 6)` should be folded into a `relu6`
operation.

When `relu6` was introduced in [this
PR](https://github.com/tenstorrent/tt-mlir/pull/5177), it wasn't fused
with convs that will be converted into matmuls as that was not supported
in tt-metal yet. That should be removed now as the problem in tt-metal
was fixed in the meantime.

### What's changed
- Added canonicalization patterns for `ClampScalarOp`
- Removed restriction for convs that can have relu6 as fused activation
in `TTNNConv2dWithActivation`

### Checklist
- [x] New/Existing tests provide coverage for changes