[ttnn.jit] Infer Worker Grids for JIT-ed Ops on DRAM Interleaved Tensors (#5433)
### Ticket
[#5279](https://github.com/tenstorrent/tt-mlir/issues/5279)

### Problem description
In D2M core, the worker grid of an op is set tp the sharding grid of its
output tensor. DRAM interleaved tensors are encoded as having a 1x1
sharding grid. This inadvertently limits ops with DRAM interleaved
output to only using a single core.

There is no easy mechanism for TTNN JIT to indicate a worker grid to the
D2M compiler on a per-op basis, so it is important for the compiler to
pick a default better than just single core.

**Testing this change revealed that block sharded tensors on non-square
grids were completely mishandled in TTNN JIT and always resulted in
corrupted output tensors**

### What's changed
- In D2MGridSelection, match `d2m.generic` ops with TTNN DRAM operands
and infer a worker grid for them as if the tensors were metal sharded
tensors:
- The `ttnn_metal_layout_cast` op is inserted as before in TTIRToD2M,
but now additional `stream_layout` and `d2m.empty()` ops are required.
- Once the grid is inferred, the layout is reblocked and an indexing map
is inferred using the `stream_layout` op
- The `d2m.generic` is rewritten to use the `stream_layout` outputs as
its operands

Previously:
```
%metal_tensor = ttir.ttnn_metal_layout_cast %ttnn_tensor : 
    tensor<1x1xHxW, #ttnn_dram_layout> 
    -> tensor<1x1xHxW, #metal_dram_layout> 
%out = d2m.generic grid  {... #ttcore.grid<1x1> ...}
    outs(%metal_tensor)
```

Now: 
```
%metal_tensor = ttir.ttnn_metal_layout_cast %ttnn_tensor : 
    tensor<1x1xHxW, #ttnn_dram_layout> 
    -> tensor<1x1xHxW, #metal_dram_layout> 
%mock_storage = d2m.empty() : tensor<GHxGWx(H/GH)x(W/GW), #metal_l1_layout>
%reblocked_metal_tensor = "d2m.stream_layout"(%metal_tensor, % mock_storage) : 
    (tensor<1x1xHxWx, #metal_dram_layout>,  tensor<GHxGWx(H/GH)x(W/GW)>, #metal_l1_layout>) 
    -> tensor<GHxGWx(H/GH)x(W/GW)>, #metal_dram_layout_with_reblock_map>
%out = d2m.generic grid  {... #ttcore.grid<GHxGW> ...}
    outs(%reblocked_metal_tensor)
```

- Output streaming only works for parallel iterator types. Updated the
verifier for `d2m.generic` to enforce this
  - https://github.com/tenstorrent/tt-mlir/issues/5446
- Modified D2MToTTNN to erase the new `stream_layout` and
`ttnn_metal_layout_cast` op structures
- Fixed bug where `stream_layout` op was incorrectly marked as legal in
the dialect conversion. The pass only worked due to the specific code
arrangement before and the stream layout rewriter was never invoked
- **Fixed bug where the TTNN grid was interpreted incorrectly in JIT
frontend and D2MToTTNN**
- When using TTNN's `core_range`, the coordinates are specified as
`(Width, Height)`. The compiler was incorrectly interpreting this as
`(Height, Width)`
- **This fixed block sharding on non-square grids which previously did
not work at all**
- Significantly expanded the coverage of layout tests
- For block sharded layouts, we now test every possible grid up to
(8x8), with varying number of tiles per core
- Discovered a bug for several grid shapes
https://github.com/tenstorrent/tt-mlir/issues/5415
- For DRAM interleaved we now test high aspect ratio shapes in addition
to square and near square shapes
- Updated IR tests

### Checklist
- [X] New/Existing tests provide coverage for changes