TTIR convoution with batch count greater than 1. (#5401)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/5307

### Problem description
`TTIR.convolution` is not able to be translated if batch group count is
greater than one.

### What's changed
Added support for this case. Inputs are sliced along batch dimension and
multiple convolutions with batch group count of exactly one are created.
This was done analogous to [stableHLO
description](https://openxla.org/stablehlo/spec#convolution).

### Checklist
- [x] New/Existing tests provide coverage for changes