Adding load tensor wrapper to adjust tensor spec after loading it (#5515)
### Ticket
N/A

### Problem description
Currently, when we load the tensor from a flatbuffer in the codegen
path, we don't guarantee that the forward function arguments are in a
proper tensor spec.

### What's changed
Added load tensor wrapper helper functions in utils that convert the
loaded tensor layout to a required spec. This is currently a bit hacky,
but as soon as we extend the EmitPy dialect with the following:
- https://github.com/tenstorrent/tt-mlir/issues/5494

We will be able to remove the utility function and generate the function
in the main file. Also, I opened this issue on the Metal side to extend
their load tensor API:
https://github.com/tenstorrent/tt-metal/issues/31131

### Checklist
- [ ] New/Existing tests provide coverage for changes