Fix CPU-hoisted `ttir.conv2d` (#5529)
### Ticket
#5495

### Problem description
CPU-hoisted `ttir.conv2d` fails because `memref.copy` gets lowered into
an invocation of MLIR runtime's `@memrefCopy` function, since it
operates on a strided memref.

We want to avoid dynamically linking the MLIR runtime libs, so the
alternative is to lower the problematic `memref.copy` to
`linalg.generic`, thus preventing it to get lowered to `@memrefCopy`
invocation.

Another issue with hoisted variant of this op is the fact that
`tensor.extract_slice` op, which can be created when lowering
`ttir.conv2d` to `linalg`, isn't a DPS op, so additional `memref.copy`
is required in order to make it CPU-hoistable.

### What's changed
- Implemented `NonContiguousMemrefCopyToLinalg` pass
- Enabled copying the `tensor.extract_slice` into the output buffer
- Un-quarantined the test

### Checklist
- [x] New/Existing tests provide coverage for changes