[stablehlo] Add transpose op support in StableHLO builder
**Description**

Add transpose op API to the stablehlo builder. And, also add the
golden reference for the transpose op.

**Motivation**

Conversion pipeline already has pattern to convert the StableHLO
transpose op to TTIR, but there is no builder.transpose method in
stablehlo_builder.py to create a stablehlo.transpose operation.

**Testing**

Added tests for transpose op in test_stablehlo_ops.

**Notes**

For ttmetal transpose dimensions > 2, skipping the tests, as they
are are not currently supported in TTIRToD2M conversion pass.