Adding pipeline option to disable workaround pass (#5667)
### Ticket
Closes https://github.com/tenstorrent/tt-mlir/issues/5648

### Problem description
We didn't have the option to disable workaround passes in the
ttir-to-ttnn pipeline. It is useful for a developer to write tests
without any workarounds applied and to mark those tests with xfail.

### What's changed
Introduced a pipeline option to enable a workaround pass. This option
defaults to true but can be set to false in builder tests. One builder
test was added to demonstrate this usage.

### Checklist
- [x] New/Existing tests provide coverage for changes