D2M: Add init hoisting pass (#4074)
### Problem description
As an optimization, we would like to hoist `*_init_*` calls as far out
of loops as possible, so that they do not get called redundantly, but
also do not conflict with one another.

### What's changed
This is a simple first cut pass which will first walk a kernel, and
build a mapping of loops with the inits that they contain.
It will walk a second time, and use the above mapping to determine which
loop nest it can safely lift each init into.
For now, it takes the conservative approach and assumes all inits
conflict with each other, so will not lift into any scope where that is
possible.

### Checklist
- [ ] New/Existing tests provide coverage for changes - note, need to
add a test for this/update existing tests