rewrites for tt-explorer content, architecture page needs more work (#5504)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-forge/issues/506)

### Problem description
TT-Explorer content is outdated and needs to be refreshed/rewritten.

### What's changed
Describe the approach used to solve the problem.
Tried out TT-Explorer, spoke with the engineer managing it - Brata - and
started a rewrite.

To note - the section about front ends isn't up-to-date completely due
to changes where TT-XLA is the new primary front end. Previously
TT-Explorer worked best with TT-Forge-FE. Next, there are still a number
of bugs in TT-Explorer. So, where installation content can be tested
thoroughly and frequently to keep it stable, this will take some ironing
out. This first pass at the content improves everything, and the
expectation should be that we will need to come back and update it
further as bugs are sorted and issues uncovered.

### Checklist
- [ ] New/Existing tests provide coverage for changes