[ttnn-jit] Using graph capture to produce IR (#5630)
### Problem description
Need to use graph capture to generate IR in `ttnn-jit`. The current
method uses python's `AST` library which is neither scalable nor
efficient. We also can't support `for` loops and `if` statements to be
`jit`ed.

### What's changed
In order to do this, we wrap the call to the python function in
`begin_graph_capture` and `end_graph_capture`. This will produce a graph
similar to the following picture for this `cosh` function:
```python
def cosh(input_tensor):
    e_pos_x = ttnn.exp(input_tensor)
    e_neg_x = ttnn.exp(ttnn.neg(input_tensor))
    nr_term = ttnn.add(e_pos_x, e_neg_x)
    output = ttnn.multiply(nr_term, 0.5)
    return output
```

<img width="4643" height="2581" alt="image"
src="https://github.com/user-attachments/assets/4b48d93c-d465-4956-8aeb-27523af2ea0e"
/>

The graph produced by `graph_capture` is complicated and unnecessarily
large. Therefore, we simplify this graph in multiple steps to achieve a
graph similar to this:
<img width="334" height="420" alt="Screenshot 2025-10-31 at 10 54 35 AM"
src="https://github.com/user-attachments/assets/4e3087a3-9ff4-4998-ab5f-873d83296e3b"
/>
Then we produce the IR from this simplified graph.

### Notes
- I ran all tests in `./test/ttnn-jit/test_eltwise.py` and confirmed
that the IRs produced by this approach and the previous one are
identical.
- I also ran all tests in `./test/ttnn-jit/test_layout.py` and confirmed
the IRs are identical (the tests are skipped right now due to some
issues).
- Using this new approach we can now decorate the functions with `for`
loops and `if` statements with `jit` (which we couldn't before). For
instance:
```python
def if_inside_for_multiple_branches(input_tensor, iterations=3):
    output = input_tensor
    for i in range(iterations):
        if i == 0:
            output = ttnn.exp(output)
        elif i == 1:
            output = ttnn.sin(output)
        else:
            output = ttnn.cos(output)
    return output

```
generates the following IR when decorated with `jit` (with
`iterations=4`):
```mlir
#l1 = #ttnn.buffer_type<l1>
#ttnn_layout = #ttnn.ttnn_layout<(d0, d1) -> (d0, d1), <1x1>, memref<2x2x!ttcore.tile<32x32, bf16>, #l1>, <block_sharded>>
module {
  func.func @op_wrapper(%arg0: tensor<64x64xbf16, #ttnn_layout>) -> tensor<64x64xbf16, #ttnn_layout> {
    %0 = "ttnn.get_device"() <{mesh_offset = #ttnn<mesh_offset 0x0>, mesh_shape = #ttnn<mesh_shape 1x1>}> : () -> !ttnn.device
    %1 = "ttnn.exp"(%arg0) {ttnn.hoist_generic_via_d2m} : (tensor<64x64xbf16, #ttnn_layout>) -> tensor<64x64xbf16, #ttnn_layout>
    %2 = "ttnn.sin"(%1) {ttnn.hoist_generic_via_d2m} : (tensor<64x64xbf16, #ttnn_layout>) -> tensor<64x64xbf16, #ttnn_layout>
    %3 = "ttnn.cos"(%2) {ttnn.hoist_generic_via_d2m} : (tensor<64x64xbf16, #ttnn_layout>) -> tensor<64x64xbf16, #ttnn_layout>
    %4 = "ttnn.cos"(%3) {ttnn.hoist_generic_via_d2m} : (tensor<64x64xbf16, #ttnn_layout>) -> tensor<64x64xbf16, #ttnn_layout>
    return %4 : tensor<64x64xbf16, #ttnn_layout>
  }
}
```
- Added tests in `test_eltwise` to test `for` loops and `if` statements.
- We also decided to maintain both `graph` based and `ast` based
approaches for IR generation. In order to do that, a flag was added to
`jit` decorator: `use_ast_compiler` which is set to `False` by default.
I also parametrized some tests to test both approaches for IR
generation.

### New Files
- `graph_trace_compiler.py`: includes all code for graph clusterization
and simplification as well as IR generation logic; the main body of this
PR.
- `ir_generator.py`: wraps both approaches for IR generation and exposes
a convenient function `generate_ir` which generates IR based on
`use_ast_compiler` flag (defaulted to `False`).

### Disclaimer
_Some code that performs clusterization and simplification is written by
@svuckovicTT found
[here](https://github.com/tenstorrent/tt-metal/compare/main...svuckovic/graph_trace#diff-298626e7e892bf7e09a05129f81b7a36edcafad89e83ba8df25451aa7ac9c14cR1)._

### Checklist
- [X] New/Existing tests provide coverage for changes

---------

Co-authored-by: Vincent Tang <157417475+vtangTT@users.noreply.github.com>
Co-authored-by: Zoe Carver <zcarver@tenstorrent.com>
Co-authored-by: Boyana Norris <bnorris@tenstorrent.com>
Co-authored-by: Vladimir Milosevic <157983820+vmilosevic@users.noreply.github.com>
Co-authored-by: jameszianxuTT <197659654+jameszianxuTT@users.noreply.github.com>
Co-authored-by: James Xu <jameszianxu@tenstorrent.com>
Co-authored-by: Radenko Pavlovic <133032400+rpavlovicTT@users.noreply.github.com>
Co-authored-by: Sanja Djukic <sdjukic@tenstorrent.com>
Co-authored-by: Muhammad Asif Manzoor <mmanzoor@tenstorrent.com>