[D2M] Insert dummy generic if func returns view (#5699)
### Ticket
NA

### Problem description
D2M views are only materialized by generics, but if a view is returned
from a func, no generic consumes it. We need a way to properly return
results of views.

### What's changed
Tiny new pass to find any case where a func is returning an
unmaterialized view and insert a generic op to materialize it (and
return the result of that instead).

### Checklist
- [ ] New/Existing tests provide coverage for changes