ttnn.rotary_embedding op + fusing (#5623)
This PR add `ttnn.rotary_embedding` op and fusing pattern for fusing
this op.

Initially we added `ttnn.rotary_embedding_llama` op with hope that we
are going to fuse RoPE into it, but it turns out that kernel
implementation of that op supports only specific case of RoPE where pair
wise rotation is done. `ttnn.rotary_embedding` peforms half rotation
which is exactly what are we getting from current llama models that we
are running.

RoPe formula in simple form looks like this:
`x * cos_cache + rotate_half(x) * sin_cache`

where x is input embedding and cos/sin cache are precomputed cos and sin
matrices which execute the rotation. `rotate_half` function takes input
embedding and split is in half at last dimension negates second half and
concatenates negated second half and first half.

Tested that I manage to test:

* EmitPy - generation is working I also tested the code itself and it's
working
* EmitC - working.  Tested using ttrt
* Tested using xla llama and tt-forge benchmarks. I can see the op is
fused and PCC is good.
* Tested quen and llama isolated `apply_rotary_emb`

The thing that doesn't work as expected (look builder test) is that
output shape of rope when we are in decode mode is wrong (more here
https://github.com/tenstorrent/tt-mlir/issues/5621). This doesn't affect
the output somehow but we should still fix it using decomposition until
we get some from ttnn team. This is only problem when rope result is
output of function and we try to copy it to host buffer, is where there
is discrepancy between expected shape and ttnn tensor shape.

Sorry for this large PR but we wanted to get things out as fast as
possible.

closes #5535