[EIO] Add Commute Patterns for Slice with Reshape (#5713)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-mlir/issues/5654)
[Link to Github
Issue](https://github.com/tenstorrent/tt-mlir/issues/4731)

### Problem description
In yolov9 model there is a pattern:
```
                input <1, 1, NHW, C>
                        |
        reshape [1, 1, NHW, C] -> [N, H, W, C]
           /                            \
slice [:, :, :, :C/2]               slice [:, :, :, C/2:]
          |                              |
reshape -> [1, 1, NHW, C/2]         reshape -> [1, 1, NHW, C/2]
          |                              |
  out <1, 1, NHW, C/2>              out <1, 1, NHW, C/2>
```
Since reshapes are not affecting slice ranges, it should be rewritten
as:

```
             input <1, 1, NHW, C>
           /                      \
slice [:, :, :, :C/2]         slice [:, :, :, C/2:]
          |                         |
  out <1, 1, NHW, C/2>        out <1, 1, NHW, C/2>
```


### What's changed
Reshape can commute with slice if it is not affecting set of elements
slice contains with and without that reshape.

The is the case when slicing is done only on dims lower than the lowest
dim changed by reshape, which we define in code as **if the leftmost
slicing dim is right to the rightmost reshape dim**. For example, these
cannot commute:
```
reshape [a,  b, c,  d]                  
     -> [a', b, c', d]                slice [:, :x, :, y:] 
         _______/                                 \_______      
  rightmost reshape dim                    leftmost slicing dim
```
When commuting we need to update slice args - begins, ends and steps:
- _left to the leftmost slicing dim_ they all have default values : 
  - `begin = 0`
  - `end = dimSize`
  - `step = 1`
so we only need to update `end` to new `dimSize`
- other dims are also _right to the rightmost reshape dim_ so they are
not affected by reshape and keep original values.

Because reshape can change number of dimensions, we use
**right-to-left** indexes as invariants, which is why they are called
`rigtmostReshapeDimRTL` and `leftmostSlicingDimRTL` in code.

This can be generalized by defining possible slicing dims as ones where
cumulative products of dimensions match before and after reshape, for
example `reshape [2,4,5,2,4] -> [8,5,2,2,2]` can commute with slice on
dim 2, making slice on dim 1, which is not covered by this PR as we
don't have need for this yet.

### Checklist
- [x] New/Existing tests provide coverage for changes