implement TODO for xorshift RNGs in test utils (#5816)
A desirable property for randomized testing utils in `"testing/Utils.h"`
is to produce random data sequences reproducible as much as possible
across OSes and stdlib versions when given only the starting seed. The
std random bit engines in std <random> may or may not have this
property, since the underlying algorithms are subject to standard spec
changes. Additionally, those std engines that have good statistical
properties also tend to be slow for drawing samples and/or have large
states.

[Marsalgia's xorshift scheme](https://en.wikipedia.org/wiki/Xorshift) is
a tried-and-true RNG that is good statistically, uses only bitwise ops,
and fits its state into a single register. This PR implements a TODO in
Utils.h to switch to it.

### Checklist
- [x] the test suite in TestAllocation automatically picks this up