Uplift third_party/tt-metal to 094a4ed7d519a3d099831c0345ffa2a6d13f5b94 2025-11-06 (#5744)
This PR uplifts the third_party/tt-metal to the
094a4ed7d519a3d099831c0345ffa2a6d13f5b94

### Note

This uplift notably creates a workaround for a bug introduced by
[this](https://github.com/tenstorrent/tt-metal/commit/f9dc6d2a9550375122c667fb04b417dc60d988cc)
metal commit. The new behavior is now that `ttnn.linear` is _always_
lowered to a matmul + add to avoid triggering the new DRAM interleaving
in metal with bad bias shapes. This should be treated as a regression.

This uplift necesarrily touches a lot of tests with the new temporary
fusion behavior (always decompose until a better solution to the bias
mismatch of the new dram interleaved linear ops can be implemented)

This means that a good number of the tests in this change must be either
reimplemented or have their changes reverted. To this end, all the test
changes have been collected in the last commit of this PR. An issue to
track these general test changes can be found here: #5789 . Please
create any sub-issues as you see fit

### Checklist
- **Frontend CI passing links**
- [x]
[tt-forge-fe](https://github.com/tenstorrent/tt-forge-fe/actions/workflows/on-pr.yml):
https://github.com/tenstorrent/tt-forge-fe/actions/runs/19312424550
- [x] [tt-xla (basic +
models)](https://github.com/tenstorrent/tt-xla/actions/workflows/manual-test.yml):
https://github.com/tenstorrent/tt-xla/actions/runs/19312426586
- **Follow-up Actions**
  - [x] **Issues filed** to follow up on incomplete changes (if any):
  - [x] **Frontend fix PRs** ready (if needed by this uplift):

---------

Co-authored-by: ctodTT <187330486+ctodTT@users.noreply.github.com>
Co-authored-by: Collin Tod <collintod@tenstorrent.com>