GridSelection for row major tensors (#5842)
### What's changed
Refactor out the physical shape calculation when updating the generic
and streamlayout ops. When calculating the physical shape to determine
the optimal grid in GridSelection, treat row majors as tilized to avoid
small shards. Additionally, loosen the assertion in updateStreamLayouts
to enable generics to use streams within their own regions.