[ttnn.jit] Remove Manual Set Seed and Xfail multiply accumulate Tests (#5874)
### Problem description
There is a `torch.manual_seed(0)` in the `create_dram_tensor` and
`create_sharded_tile_tensor` helper functions in `utils.py.` This means
tests in ttnn-jit are being run with identical inputs. For example.
mul_add is computing `a + a * a` instead of `a + b * c` where a, b, and
c are tensors with different values.

### What's changed
Removed the set manual seed from helper functions. This revealed failing
allclose checks for multiply-accumulate (mul_add) which are now marked
with xfail. Issue tracking here #5873

### Checklist
- [x] New/Existing tests provide coverage for changes