Model `ttnn.paged_scaled_dot_product_attention_decode` op (#5805)
### Ticket
Closes
https://github.com/tenstorrent/tt-mlir/issues/4214?issue=tenstorrent%7Ctt-mlir%7C4264

### Problem description
We do not model `ttnn.paged_scaled_dot_product_attention_decode`

NOTE: There is an issue which causes `PagedUpdateCacheOp`
silicon/runtime tests to hang. It seems to be because we cannot choose
which integers are actually placed in the page table tensor, which is
used to index the cache. Generating random values for the page table
will not suffice as we do not know what happens if the indices in the
tensor are out of bounds (likely causing hang). This op also indexes its
inputs using a page table tensor. And so, we do not test runtime in
`TestOpModelInterface.cpp` or have a silicon test:
https://github.com/tenstorrent/tt-mlir/issues/5738

### What's changed
- Model `PagedScaledDotProductAttentionDecode` in runtime, ttnn, ttir,
and shlo -> ttir as custom call.
- Requires workaround forcing the `pageTable` and `curPosTensor` args to
be in ROW MAJOR.
- Add OpModelLib and OpModelInterface tests
- Also removed a verification invariant for `PagedUpdateCache` which
makes sure `numHeads == 32` as this does not appear to be necessary from
testing I've done.
- Had to include `NOLINT` barriers for a specific linter error caused by
the use/passing around of an optional `APFloat` this was investigated
heavily when bringing up `ScaledDotProductAttentionDecode` and we
believe it is a false-negative.
- This same `NOLINT` is included in some areas of
`ScaledDotProductAttentionDecode` currently.
### Checklist
- [X] New/Existing tests provide coverage for changes