Support lowering partially indexed DMAs (#5788)
This commit handles a missing lowering case for DMA's where the DMA is
partially indexed and does not have the numElems attribute set. This
implicitly means, gather all of the indices that aren't applied (similar
to python slice semantics).

This is implemented by:
- Adding a trivial loop around the unapplied indices over the remaining
shape.
- Disambiguating this DMA form by introducing `hasExplicitSize`. If
false it means we need to lower via bullet 1, if true, it means some
other lowering proved the unapplied indices are contiguous i.e. via
affine lowering.

As an example, the following dma:
```mlir
%src : memref<1x2x2x2x!ttcore.tile<32x32, f32>, #ttcore.shard<16384x4096>, #l1_>
%dst :  memref<2x2x!ttcore.tile<32x32, f32>, #l1_>
%0 = d2m.dma %src[%c0, %c1], %dst : !d2m.mem_tx
```

Becomes:
```mlir
%2 = scf.for %arg1 = 0 to 2 step 1 iter_args(%arg2 = %1) -> (!d2m.mem_tx) {
  %3 = scf.for %arg3 = 0 to 2 step 1 iter_args(%arg4 = %arg2) -> (!d2m.mem_tx) {
    %tx = d2m.dma %src [%c0, %c1, %arg1, %arg3], %dst [%arg1, %arg3], <1> : !d2m.mem_tx
    scf.yield %tx : !d2m.mem_tx
  }
  scf.yield %3 : !d2m.mem_tx
}
```