Docker build in the CI refactor (#5885)
### Ticket
/

### Problem description
Execute the docker build job in the CI on CIv1 builder machines labelled
with `docker-builder`.

### What's changed
The docker build job in the CI is now executed on the CIv1 builder
machines.
Two changes were required to enable this:
- Set the permissions of the workspace to the current user to avoid the
permisisons issue
- Limit the cmake's parallel level during the docker image building to
avoid OOM error which is caused when the toolchain is built without the
parallel level limiting

### Checklist
- [ ] Docker build job passing with the changes from this PR ->
[job](https://github.com/tenstorrent/tt-mlir/actions/runs/19425546544/job/55572082145)