[Uplift Enhancement] Permit custom metal commit hash for manual uplift, with safety check to disallow downgrade + metal branch commits (#5718)
### Ticket
None

### Problem description

Not much flexibility to manual uplift job. We might want to uplift a
smaller section of commits from metal, but we can't since the job
selects the most recent commit.

### What's changed
- Add optional custom metal commit (default nothing = use tip of metal
main like before)
- Add safety checks for custom metal commit -> Disallow uplifting a
metal downgrade, disallow uplifting to a non-main metal branch

This is a pure QOL change for uplifters. It could have been done
manually before by manually editing a commit post-creation

This is an additional enhancement on top of #5616 

### Checklist
- [x] New/Existing tests provide coverage for changes (see comments for
validation)