Move rope fusing into ttir (#5956)
It seems that rope fusing got broken when running whole llama. This is
because erase inverse ops inserts some TMs in the middle of the pattern.
To fix this I moved this pattern to TTIR temporarily until better
solution is found. The only new addition wrt last fusing pattern in TTNN
is that when matching unsqeezed cos/sin we are checking if there is
broadcast op in between reshape and multiply. If there is we will take
broadcast input instead of broadcast because those dimensions are
compatible with ttnn.