Fix Scheduler for ops with multiple outputs (#5977)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-mlir/issues/5884)

### Problem description
When running models that include ops with multiple outputs, such as
split_query_key_value_and_split_heads, and the
memory-layout-analysis-enabled flag is enabled, ops may be incorrectly
sorted, with users of a multi-output op appearing before the op itself.

This occurs because the scheduler assumes ops have only one output, so
it considers only the users of the first result.
### What's changed
Fixed dependency initialization in the scheduler so that all results are
processed.

### Checklist
- [x] New/Existing tests provide coverage for changes